Support commands, fix PM reply bug (BREAKING CHANGE)
The new feature is that you can now register for a command, directly, and
the bot will take care of everything.  The bugfix is that previously if you
replied to a PM, the message would be sent back to the bot, not to the user
who sent it.  The breaking change is that I renamed the `#source` method on
`Message` to `#channel`, because the old name sucked.